services/orchestrator/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client
httpx==0.25.2

duckduckgo_search==6.1.3

# Data validation
pydantic==2.5.0

# Environment and config
python-dotenv==1.0.0
pyyaml==6.0.1

# Audio processing and speech recognition
SpeechRecognition==3.10.0
pydub==0.25.1

# LLM and embeddings
numpy==1.24.3
faiss-cpu==1.7.4

# CORS middleware
python-multipart==0.0.6

# Additional utilities